moved select test
