+++ src/js/components/Form/__tests__/Form-test-controlled.js
@@ -8,7 +8,6 @@ import { Form } from '..';
 import { FormField } from '../../FormField';
 import { Button } from '../../Button';
 import { TextInput } from '../../TextInput';
-import { Select } from '../../Select';
 
 describe('Form controlled', () => {
   afterEach(cleanup);
@@ -257,32 +256,5 @@ describe('Form controlled', () => {
     fireEvent.click(getByText('Reset'));
     expect(onReset).toBeCalledTimes(1);
     expect(queryByText('Input has changed')).toBeNull();
-  });
-
-  test('form with select', () => {
-    const Test = () => {
-      const [value, setValue] = React.useState('medium');
-      return (
-        <Grommet>
-          <Form>
-            <FormField>
-              <Select
-                a11yTitle="select form"
-                name="select"
-                placeholder="test input"
-                options={['small', 'medium', 'large']}
-                value={value}
-                onChange={({ option }) => setValue(option)}
-              />
-            </FormField>
-          </Form>
-        </Grommet>
-      );
-    };
-    const { getByPlaceholderText, container } = render(<Test />);
-    fireEvent.change(getByPlaceholderText('test input'), {
-      target: { value: 'small' },
-    });
-    expect(container.firstChild).toMatchSnapshot();
   });
 });
